kobject: Buffer uevents delivered across suspend
If userspace is waiting on a netlink socket with EPOLLWAKEUP (Android
does this) and a driver wants to emit a uevent during suspend (Intel
machinecheck does this), the delivery of the skb will activate a
wakeup source synchronously and abort the suspend, even if userspace
has already been frozen.

Simply buffer events across suspend and deliver them on wakeup.

Issue: APDEV-834
Change-Id: Id705dc790eaa07651cce1e807ccbbb39e27d869b
Signed-off-by: Andy Ross <andrew.j.ross@intel.com>
Reviewed-on: https://android.intel.com/160368
Tested-by: sys_abtbuild <sys_abtbuild@intel.com>
Reviewed-by: Zhang, Jianxun <jianxun.zhang@intel.com>
Reviewed-by: Holmberg, Hans <hans.holmberg@intel.com>
Reviewed-by: Phillips, Scott D <scott.d.phillips@intel.com>
Reviewed-by: Jovanovic, Radivoje <radivoje.jovanovic@intel.com>
Tested-by: Jovanovic, Radivoje <radivoje.jovanovic@intel.com>
Reviewed-by: Leung, Daniel <daniel.leung@intel.com>