FROM_UPSTREAM iio: hid-sensors: Fix power and report state
Take subset as backport. partial changes are already there.

Author: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
Date:   Wed Nov 27 22:19:00 2013 +000

In the original HID sensor hub firmwares all Named array enums were
to 0-based. But the most recent hub implemented as 1-based,
because of the implementation by one of the major OS vendor.
Using logical minimum for the field as the base of enum. So we add
logical minimum to the selector values before setting those fields.

Some sensor hub FWs already changed logical minimum from 0 to 1
to reflect this and hope every other vendor will follow.
There is no easy way to add a common HID quirk for NAry elements,
even if the standard specifies these field as NAry, the collection
used to describe selectors is still just "logical".

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
Signed-off-by: Jonathan Cameron <jic23@kernel.org>

For: AIAF-4024

Upstream-Commit-Id: 751d17e23a9f7c8e0bca5c0b2e8d39af655ecd2a
Change-Id: I7d1477a6e6711580a928a564de1e28b12cb417c6
Signed-off-by: Archana Patni <archana.patni@intel.com>
Reviewed-on: https://android.intel.com/160117
Reviewed-by: Sesha, Subramony <subramony.sesha@intel.com>
Reviewed-by: Nordell, Joakim <joakim.nordell@intel.com>
Tested-by: sys_abtbuild <sys_abtbuild@intel.com>
Reviewed-by: Purdila, Octavian <octavian.purdila@intel.com>
Reviewed-by: Odeberg, Hans <hans.odeberg@intel.com>