FROM_UPSTREAM: HID: hid-sensor-hub:Fix buggy report descriptors
This addresses regression caused by commit id
"751d17e23a9f7"  iio: hid-sensors: Fix power and
report state. This commit removed a quirk, to change
the enumeration base to 1 from 0 based on an CONFIG
parameter. There was objection to add more changes
under this quirk, instead suggested to add an HID
quirk. But there is no easy way to add HID quirk as
the reports are not properly using collection class.

The solution was to use logical minimum, which is a correct way.
There were changes done in firmware to address this.

Unfortunately some devices, still use old FW and can't be upgraded
to newer version on Linux devices as there is no FW upgrade tool
available for Linux devices. So we need to fix report descriptors,
for such devices. This will not have any impact, if the FW uses
logical 1 as minimum.

In this patch we look for usage id for "power and report state", and
modify logical minimum value to 1.

Background on enum:
In the original HID sensor hub firmwares all Named array enums were
to 0-based. But the most recent hub implemented as 1-based,
because of the implementation by one of the major OS vendor.
Using logical minimum for the field as the base of enum. So we add
logical minimum to the selector values before setting those fields.
Some sensor hub FWs already changed logical minimum from 0 to 1
to reflect this and hope every other vendor will follow.

There is no easy way to add a common HID quirk for NAry elements,
even if the standard specifies these field as NAry, the collection
used to describe selectors is still just "logical".

Author: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
Date:   Tue Jan 14 09:30:20 2014 -0800

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>

For: AIAF-4024

Upstream-Commit-Id: 875e36f8a2ce6cfc2010af6c91fb5351bd530d48
Change-Id: I6412d4d930d89cec8841d570ed87d400b298a73b
Signed-off-by: Archana Patni <archana.patni@intel.com>
Reviewed-on: https://android.intel.com/160119
Reviewed-by: Sesha, Subramony <subramony.sesha@intel.com>
Tested-by: Sesha, Subramony <subramony.sesha@intel.com>
Tested-by: sys_abtbuild <sys_abtbuild@intel.com>
Reviewed-by: Purdila, Octavian <octavian.purdila@intel.com>
Reviewed-by: Odeberg, Hans <hans.odeberg@intel.com>